Refactor example_pathplanning
